NateN-DreamInterpreter/requirements.txt
flask==2.3.3
google-generativeai==0.3.2
PyPDF2==3.0.1
python-dotenv==1.0.0
sentence-transformers==2.2.2
torch==2.1.0